+++ src/components/Wizard/ModelSelection/CesiumGeojsonFootprint.tsx
@@ -1,13 +1,22 @@
-import { useEffect, useState } from "react";
-import { CesiumGeojsonLayer, RCesiumGeojsonLayerProps } from "@map-colonies/react-components";
-import { FeatureCollection } from "geojson";
-import { getLayerFootprint } from "../../../utils/Cesium/GetLayerFootprint";
-import { FlyTo } from "../../../utils/Cesium/FlyTo";
+import { useEffect, useState } from 'react';
+import { FeatureCollection } from 'geojson';
+import {
+  CesiumColor,
+  CesiumConstantProperty,
+  CesiumGeojsonLayer,
+  CesiumRectangle,
+  RCesiumGeojsonLayerProps
+} from '@map-colonies/react-components';
+import { FlyTo } from '../../../utils/Cesium/FlyTo';
+import { getLayerFootprint } from '../../../utils/Cesium/GetLayerFootprint';
+import { FOOTPRINT_BORDER_WIDTH } from '../../../utils/Const';
 
 interface CesiumGeojsonFootprintProps extends RCesiumGeojsonLayerProps {
   id: string;
   setFinishedFlying?: (finished: boolean) => void;
 }
+
+const FOOTPRINT_BORDER_COLOR = CesiumColor.DODGERBLUE;
 
 export const CesiumGeojsonFootprint: React.FC<RCesiumGeojsonLayerProps & CesiumGeojsonFootprintProps> = (props) => {
   const [layersFootprints, setlayersFootprints] = useState<FeatureCollection>();
@@ -36,6 +45,24 @@ export const CesiumGeojsonFootprint: React.FC<RCesiumGeojsonLayerProps & CesiumG
       <CesiumGeojsonLayer
         {...props}
         data={layersFootprints}
+        onLoad={(geoJsonDataSouce): void => {
+          geoJsonDataSouce.entities.values.forEach(item => {
+            if (item.polyline) {
+              (item.polyline.width as CesiumConstantProperty).setValue(FOOTPRINT_BORDER_WIDTH);
+              // typings issue in CESIUM for reference https://github.com/CesiumGS/cesium/issues/8898
+              // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+              // @ts-ignore
+              item.polyline.material = FOOTPRINT_BORDER_COLOR;
+            }
+            if (item.polygon) {
+              (item.polygon.outlineColor as CesiumConstantProperty).setValue(FOOTPRINT_BORDER_COLOR);
+              // typings issue in CESIUM for reference https://github.com/CesiumGS/cesium/issues/8898
+              // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+              // @ts-ignore
+              item.polygon.material = CesiumColor.fromRandom({ alpha: 0.4 });
+            }
+          });
+        }}
       />
       {
         flyTo &&
@@ -47,4 +74,4 @@ export const CesiumGeojsonFootprint: React.FC<RCesiumGeojsonLayerProps & CesiumG
       }
     </>
   );
-}
+};